radjangoproject/requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2022.7.1
sqlparse==0.4.3
